Remove dtype to support regression tasks

+++ nodecls/datamodules/datasets.py
@@ -12,8 +12,8 @@ class NumPyDataset(Dataset):
         assert split_type in ["train", "val", "test"]
         data = np.load(data_path, allow_pickle=True)
         self.masks = data[f"{split_type}_masks"]
-        self.feats = data["features"].astype(np.float32)
-        self.labels = data["labels"].astype(np.int64)
+        self.feats = data["features"]
+        self.labels = data["labels"]
         if apply_masking:
             self.feats = self.feats[self.masks]
             self.labels = self.labels[self.masks]
